requirements.txt
requests==2.32.3
flask==3.1.0
python-dotenv==1.0.1
pandas==2.2.3
matplotlib==3.10.0
mongoengine==0.29.1
pymongo==4.10.1
pydash==8.0.4
